+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.4.0 - 2016-5-7
+
+- Support for class-names inside pseudo-selectors like: ".foo:not(.bar)"
+- Added new "output" option, aggregating "dist", "outputName" and "type" into a single option, with fallback to original options
+- Allowing for explicit/template/callback values for options: "classnameFormat", "output", "dist", "outputName", "type"
+- Code refactor
+
 # 0.3.1 - 2016-5-2
 
 Add customize output file name.[[@ofzza](https://github.com/ofzza)] [[#12](https://github.com/ctxhou/postcss-hash-classname/issues/12)]